- création du sous-namespace "accessibleModel"
- ajout des tables :
     usager
     usager_facebook
     usager_google
     probleme
     type
