+++ model/accessibleModel/Evenement.php
@@ -1,5 +1,5 @@
 <?php
-namespace  model;
+namespace model\accessibleModel;
 /**
  * Created by PhpStorm.
  * User: cedric
@@ -11,15 +11,10 @@ class Evenement extends DataAccess {
     /**
      * Evenement constructor.
      */
-    public function __construct($pdo)
-    {
+    public function __construct($pdo){
         parent::__construct($pdo);
         $this->_idColumnName = EvenementTable::COLUMNS['ID'];
+        $this->_columns = EvenementTable::COLUMNS;
     }
-
-    public function getAllColumns(){
-        return EvenementTable::getAllColumns();
-    }
-
 
 }
