+++ model/accessibleModel/RegionTable.php
@@ -1,0 +1,12 @@
+<?php
+namespace model\accessibleModel;
+
+class RegionTable {
+
+    const TABLE_NAME = "region";
+
+    const COLUMNS =
+        ["ID"=>"idRegion",
+        "NOM"=>"nom"];
+
+}
